[add] store 간 데이터 저장하는 코드 훅으로 분리
