+++ libs/hooks/src/index.ts
@@ -31,3 +31,4 @@ export * from './lib/useSocialLoginFailAlarm';
 export * from './lib/mutation/useCreateShippingGroup';
 export * from './lib/mutation/useDeleteShippingGroup';
 export * from './lib/queries/useShippingGroupList';
+export * from './lib/shipping-group-hooks';
